requirements.txt
Django==3.0.2
django-extensions==2.2.6
djangorestframework==3.11.0
